refactor team edit pages, combine similar parts to one component
